Issue #1542 This code is an attempt, to temporarily disable the SSO force redirection for debugging and testing purpose. The basic idea is, to simply add the parameter disable_redirect=1 to the query string, to disable the redirect. E.g.: /users/login?disable_redirect=1
The issue with this approach is, that the HTML template gets loaded without any input fields for the login.
The template URL is hardcoded and can not take any parameters (at least to my knowledge)
Therefore this code currently disabled the redirect, but the user has no login fields. Not that helpful :(